Added superadmin

+++ resources/views/livewire/admin/location-management/view-areas-location-management.blade.php
@@ -102,11 +102,13 @@
                                                                 <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M18.364 5.636l-12.728 12.728M12 2a10 10 0 1010 10A10 10 0 0012 2z"></path>
                                                             </svg>
                                                         </button>
-                                                        <button type="button" wire:click="deleteArea('{{ $area->id }}')" class="p-1 text-red-600 hover:text-red-800 cursor-pointer" title="Delete">
-                                                            <svg class="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                                                                <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6M9 7h6m2 0H7m2 0V5a2 2 0 012-2h2a2 2 0 012 2v2"></path>
-                                                            </svg>
-                                                        </button>
+                                                        @if (auth()->check() && (int) auth()->user()->user_type === 2)
+                                                            <button type="button" wire:click="deleteArea('{{ $area->id }}')" class="p-1 text-red-600 hover:text-red-800 cursor-pointer" title="Delete">
+                                                                <svg class="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                                                                    <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6M9 7h6m2 0H7m2 0V5a2 2 0 012-2h2a2 2 0 012 2v2"></path>
+                                                                </svg>
+                                                            </button>
+                                                        @endif
                                                     @endif
                                                 </div>
                                             </td>
